address various comments
